# store-gateway: reduce syscalls for looking up symbols