… pkg/storegateway/indexheader/index/symbols_test.go
… pkg/storegateway/series_refs_test.go