#### Context

Related to https://github.com/grafana/mimir/issues/4593#issuecomment-1550928468


A symbol in this context is a string stored in the TSDB index. Every timeseries in the TSDB index has a reference to the symbols in its labels instead of the actual strings. The symbols table (mapping from symbol reference to string) is also present in the index header that the store-gateway stores locally.

For series requests the current way of reading symbols is as follows: for each symbol (label name or label value)
* the lazy index header reader acquires a read lock on all index header readers and checks `time.Now()` ([code](https://github.com/grafana/mimir/blob/db5713d7c2fde8ac1c22f51919e43e7821060e8e/pkg/storegateway/indexheader/lazy_binary_reader.go#L168-L175))
* label names: these are already cached in a go map in memory so looking them up is cheap ([code](https://github.com/grafana/mimir/blob/baaf4d22ee29916e3a70e1a47245e56223dbf163/pkg/storegateway/indexheader/stream_binary_reader.go#L46-L48))
* label values: 
	* there is a cache of 1024 symbols per block which keeps some symbols ([code](https://github.com/grafana/mimir/blob/baaf4d22ee29916e3a70e1a47245e56223dbf163/pkg/storegateway/indexheader/stream_binary_reader.go#L51-L55))
	* on a cache miss the index header reader obtains a file reader to the index header, offsets to a offset near the requested symbol in the symbols table, reads the symbol and returns it, closes the file reader or returns it to a buffer (currently the buffer has a size of 1)

#### Opportunities for improvement

* Frequent calls to `time.Now()` and `RWMutex.RLock()/RUnlock()`; these are relative expensive to the point that time.Now accounts for 1% of CPU; the mutex is also global
* Frequent file opens; opening a file is a syscall which can be expensive
* Random file reads; symbols are read in the other they appear in the series and in the order that series appear; we end up reading from all over the symbols table when populating the symbols of series

When there are requests returning more series about 16% of CPU is spent looking up symbols for said series. See https://github.com/grafana/mimir/issues/4593#issuecomment-1550928468

#### What this PR does

It tries to address the challenges above. The flow for label names is generally the same.

For label values:

* On requests that have less than 256 series: do the same as we currently do - single file reader per label value symbol
	* See [this comment](https://github.com/grafana/mimir/pull/5073#issuecomment-1560966374) for why 256 was chosen
* On requests that have more than 256 series: 
	* open a single file reader per set/batch (5000 series at the moment)
	* create a sorted list of symbols references to look up and a map from symbol reference to string
	* look up the symbol strings in sequence to reduce random reads and populate the map
	* for each symbol reference in each series lookup in the map and get a string
	* close file reader or return to buffer

#### Benchmark

Ran on linux/amd64 with an HDD

<details>
<summary>BenchmarkOpenBlockSeriesChunkRefsSetsIterator</summary>

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/storegateway
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                                                                  │ store-gateway-clean-up-chunks-fetching-deprecate-bucketed-chunks-pool-4996-942e16cfc.txt │ Clean-up-symbolization-67b7a0af7.txt │
                                                                                                  │                                          sec/op                                          │    sec/op      vs base               │
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X"-4                                                                                                                25.49µ ± ∞ ¹    25.31µ ± ∞ ¹        ~ (p=1.000 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j="foo"-4                                                                                                        75.31µ ± ∞ ¹    97.49µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j!="foo"-4                                                                                                       105.2µ ± ∞ ¹    101.8µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"XXX|YYY"-4                                                                                                         12.48µ ± ∞ ¹    13.61µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"X.+"-4                                                                                                            12.460µ ± ∞ ¹    5.059µ ± ∞ ¹  -59.40% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"X|Y|Z"-4                                                                                                           59.72µ ± ∞ ¹    59.08µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1"-4                                                                                                               2584.3m ± ∞ ¹    417.2m ± ∞ ¹  -83.86% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j="foo"-4                                                                                                       2395.0m ± ∞ ¹    307.9m ± ∞ ¹  -87.14% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n="1"-4                                                                                                       2400.2m ± ∞ ¹    305.8m ± ∞ ¹  -87.26% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j!="foo"-4                                                                                                      2415.0m ± ∞ ¹    298.5m ± ∞ ¹  -87.64% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*"-4                                                                                                              10.951 ± ∞ ¹     8.137 ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".+"-4                                                                                                               14.11 ± ∞ ¹     11.15 ± ∞ ¹  -20.99% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"^.+$",j=~"X.+"-4                                                                                                   23.32m ± ∞ ¹    23.70m ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~""-4                                                                                                                 3.021 ± ∞ ¹     3.014 ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!=""-4                                                                                                                 13.71 ± ∞ ¹     11.35 ± ∞ ¹  -17.22% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",j="foo"-4                                                                                               2364.6m ± ∞ ¹    322.7m ± ∞ ¹  -86.35% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",j="foo"-4                                                                                              48.09µ ± ∞ ¹    48.00µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",i!="2",j="foo"-4                                                                                        2433.9m ± ∞ ¹    338.8m ± ∞ ¹  -86.08% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!=""-4                                                                                                           5.140 ± ∞ ¹     2.963 ± ∞ ¹  -42.36% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j="foo"-4                                                                                                   4.913 ± ∞ ¹     3.005 ± ∞ ¹  -38.83% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"X.+"-4                                                                                                 23.84m ± ∞ ¹    23.17m ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"XXX|YYY"-4                                                                                             33.52µ ± ∞ ¹    33.84µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                                                                                             105.5µ ± ∞ ¹    104.6µ ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",j="foo"-4                                                                                                 4.920 ± ∞ ¹     2.882 ± ∞ ¹  -41.43% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"1.+",j="foo"-4                                                                                               564.7m ± ∞ ¹    338.7m ± ∞ ¹  -40.03% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!="2",j="foo"-4                                                                                          4.931 ± ∞ ¹     2.888 ± ∞ ¹  -41.42% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4                                                                                        4.976 ± ∞ ¹     3.162 ± ∞ ¹  -36.45% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4                                                                                 47.25µ ± ∞ ¹    47.40µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"0xxx|1xxx|2xxx"-4                                                                                                  710.8µ ± ∞ ¹    679.1µ ± ∞ ¹   -4.46% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"(0|1|2)xxx"-4                                                                                                      699.7µ ± ∞ ¹    691.7µ ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"[0-2]xxx"-4                                                                                                        707.1µ ± ∞ ¹    721.5µ ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!~[0-2]xxx-4                                                                                                          11.119 ± ∞ ¹     8.666 ± ∞ ¹  -22.06% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*",_i!~[0-2]xxx-4                                                                                                 11.197 ± ∞ ¹     8.579 ± ∞ ¹  -23.38% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"<unique_prefix>.+"-4                                                                                               271.7µ ± ∞ ¹    267.9µ ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"<unique_prefix>.+"-4                                                                                         3.268m ± ∞ ¹    3.222m ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!~"<unique_prefix>.+"-4                                                                                        2560.9m ± ∞ ¹    418.5m ± ∞ ¹  -83.66% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4                                                                               32.64m ± ∞ ¹    31.48m ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".+",i=~"<unique_prefix>.+"-4                                                                               103.2m ± ∞ ¹    106.2m ± ∞ ¹   +2.91% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".*",i=~"<unique_prefix>.+"-4                                                                               28.88m ± ∞ ¹    27.92m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".*",n=~".*",i=~"<unique_prefix>.+"-4                                                                               319.2µ ± ∞ ¹    318.1µ ± ∞ ¹        ~ (p=1.000 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".+",n=~".+",i=~"<unique_prefix>.+"-4                                                                               144.8m ± ∞ ¹    141.1m ± ∞ ¹   -2.56% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/p!=""-4                                                                                                                 3.947 ± ∞ ¹     1.801 ± ∞ ¹  -54.37% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X"-4                                                                                                                   25.30µ ± ∞ ¹    25.54µ ± ∞ ¹   +0.94% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j="foo"-4                                                                                                           47.18µ ± ∞ ¹    46.33µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j!="foo"-4                                                                                                          46.60µ ± ∞ ¹    47.02µ ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"XXX|YYY"-4                                                                                                            5.300µ ± ∞ ¹    5.338µ ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"X.+"-4                                                                                                                5.195µ ± ∞ ¹    5.230µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"X|Y|Z"-4                                                                                                              58.57µ ± ∞ ¹    60.69µ ± ∞ ¹   +3.62% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1"-4                                                                                                                  2538.1m ± ∞ ¹    429.6m ± ∞ ¹  -83.08% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j="foo"-4                                                                                                          4718.2m ± ∞ ¹    315.1m ± ∞ ¹  -93.32% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n="1"-4                                                                                                          5363.1m ± ∞ ¹    305.9m ± ∞ ¹  -94.30% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j!="foo"-4                                                                                                         2650.1m ± ∞ ¹    301.3m ± ∞ ¹  -88.63% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".*"-4                                                                                                                 11.161 ± ∞ ¹     8.570 ± ∞ ¹  -23.21% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".+"-4                                                                                                                  13.56 ± ∞ ¹     17.80 ± ∞ ¹        ~ (p=1.000 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"^.+$",j=~"X.+"-4                                                                                                      23.17m ± ∞ ¹    23.43m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~""-4                                                                                                                    3.006 ± ∞ ¹     3.058 ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i!=""-4                                                                                                                    13.73 ± ∞ ¹     11.39 ± ∞ ¹  -17.04% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",j="foo"-4                                                                                                  2352.1m ± ∞ ¹    324.9m ± ∞ ¹  -86.19% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",j="foo"-4                                                                                                 46.72µ ± ∞ ¹    48.81µ ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",i!="2",j="foo"-4                                                                                           2378.9m ± ∞ ¹    336.8m ± ∞ ¹  -85.84% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!=""-4                                                                                                              5.029 ± ∞ ¹     3.010 ± ∞ ¹  -40.15% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j="foo"-4                                                                                                      4.945 ± ∞ ¹     2.883 ± ∞ ¹  -41.70% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"X.+"-4                                                                                                    22.55m ± ∞ ¹    23.17m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"XXX|YYY"-4                                                                                                34.63µ ± ∞ ¹    33.95µ ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                                                                                                106.6µ ± ∞ ¹    105.4µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",j="foo"-4                                                                                                    4.917 ± ∞ ¹     2.886 ± ∞ ¹  -41.30% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"1.+",j="foo"-4                                                                                                  571.1m ± ∞ ¹    335.9m ± ∞ ¹  -41.18% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!="2",j="foo"-4                                                                                             4.918 ± ∞ ¹     2.866 ± ∞ ¹  -41.74% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4                                                                                           4.961 ± ∞ ¹     3.141 ± ∞ ¹  -36.69% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4                                                                                    47.67µ ± ∞ ¹    47.69µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"0xxx|1xxx|2xxx"-4                                                                                                     688.6µ ± ∞ ¹    709.0µ ± ∞ ¹        ~ (p=0.057 n=4)
…