@dimitarvdimitrov 2023-05-24
##### Choosing a threshold

I ran `BenchmarkOpenBlockSeriesChunkRefsSetsIterator` comparing thresholds for using the optimization: 64 vs 256 vs 512. There is a visible degradation when running with a threshold of 64 series in the cases with `i=~"0xxx|1xxx|2xxx"`, `i=~"(0|1|2)xxx"`, and `i=~"[0-2]xxx"` as those select 150 series.

The other changes in results I think are flukes because their cases select either 0 series or multiple batches of 5000 series, so whether the threshold is 64, 256 or 512 it shouldn't matter.

<details>
<summary>`BenchmarkOpenBlockSeriesChunkRefsSetsIterator` 256 vs 64</summary>

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/storegateway
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                                                                  │ Clean-up-symbolization-67b7a0af7.txt │ Have-a-lower-bound-to-using-single-pass-stringify-bd654b059.txt │
                                                                                                  │                sec/op                │                 sec/op                  vs base                 │
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X"-4                                                            25.31µ ± ∞ ¹                             25.94µ ± ∞ ¹   +2.46% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j="foo"-4                                                    97.49µ ± ∞ ¹                             47.91µ ± ∞ ¹  -50.86% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j!="foo"-4                                                  101.84µ ± ∞ ¹                             47.71µ ± ∞ ¹  -53.15% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"XXX|YYY"-4                                                    13.610µ ± ∞ ¹                             5.595µ ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"X.+"-4                                                         5.059µ ± ∞ ¹                             5.482µ ± ∞ ¹   +8.36% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"X|Y|Z"-4                                                       59.08µ ± ∞ ¹                             58.70µ ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1"-4                                                            417.2m ± ∞ ¹                             426.3m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j="foo"-4                                                    307.9m ± ∞ ¹                             313.2m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n="1"-4                                                    305.8m ± ∞ ¹                             310.2m ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j!="foo"-4                                                   298.5m ± ∞ ¹                             303.0m ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*"-4                                                           8.137 ± ∞ ¹                              8.170 ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".+"-4                                                           11.15 ± ∞ ¹                              11.29 ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"^.+$",j=~"X.+"-4                                               23.70m ± ∞ ¹                             24.75m ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~""-4                                                             3.014 ± ∞ ¹                              3.030 ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!=""-4                                                             11.35 ± ∞ ¹                              11.26 ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",j="foo"-4                                            322.7m ± ∞ ¹                             322.6m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",j="foo"-4                                          48.00µ ± ∞ ¹                             48.48µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",i!="2",j="foo"-4                                     338.8m ± ∞ ¹                             334.7m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!=""-4                                                       2.963 ± ∞ ¹                              2.971 ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j="foo"-4                                               3.005 ± ∞ ¹                              2.867 ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"X.+"-4                                             23.17m ± ∞ ¹                             28.80m ± ∞ ¹  +24.29% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"XXX|YYY"-4                                         33.84µ ± ∞ ¹                             35.44µ ± ∞ ¹   +4.70% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                                         104.6µ ± ∞ ¹                             208.7µ ± ∞ ¹  +99.50% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",j="foo"-4                                             2.882 ± ∞ ¹                              5.600 ± ∞ ¹  +94.30% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"1.+",j="foo"-4                                           338.7m ± ∞ ¹                             341.5m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!="2",j="foo"-4                                      2.888 ± ∞ ¹                              2.940 ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4                                    3.162 ± ∞ ¹                              3.216 ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4                             47.40µ ± ∞ ¹                             49.29µ ± ∞ ¹   +3.99% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"0xxx|1xxx|2xxx"-4                                              679.1µ ± ∞ ¹                             879.5µ ± ∞ ¹  +29.52% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"(0|1|2)xxx"-4                                                  691.7µ ± ∞ ¹                             849.1µ ± ∞ ¹  +22.77% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"[0-2]xxx"-4                                                    721.5µ ± ∞ ¹                             809.1µ ± ∞ ¹  +12.13% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!~[0-2]xxx-4                                                       8.666 ± ∞ ¹                              8.401 ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*",_i!~[0-2]xxx-4                                              8.579 ± ∞ ¹                              8.517 ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"<unique_prefix>.+"-4                                           267.9µ ± ∞ ¹                             276.2µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"<unique_prefix>.+"-4                                     3.222m ± ∞ ¹                             3.436m ± ∞ ¹   +6.65% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!~"<unique_prefix>.+"-4                                     418.5m ± ∞ ¹                             436.3m ± ∞ ¹   +4.26% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4                           31.48m ± ∞ ¹                             31.99m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".+",i=~"<unique_prefix>.+"-4                           106.2m ± ∞ ¹                             107.9m ± ∞ ¹        ~ (p=0.057 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".*",i=~"<unique_prefix>.+"-4                           27.92m ± ∞ ¹                             26.05m ± ∞ ¹   -6.68% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".*",n=~".*",i=~"<unique_prefix>.+"-4                           318.1µ ± ∞ ¹                             335.5µ ± ∞ ¹   +5.50% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".+",n=~".+",i=~"<unique_prefix>.+"-4                           141.1m ± ∞ ¹                             144.7m ± ∞ ¹   +2.55% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/p!=""-4                                                             1.801 ± ∞ ¹                              1.726 ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X"-4                                                               25.54µ ± ∞ ¹                             25.15µ ± ∞ ¹   -1.53% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j="foo"-4                                                       46.33µ ± ∞ ¹                             46.98µ ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j!="foo"-4                                                      47.02µ ± ∞ ¹                             47.03µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"XXX|YYY"-4                                                        5.338µ ± ∞ ¹                             5.494µ ± ∞ ¹   +2.92% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"X.+"-4                                                            5.230µ ± ∞ ¹                             5.402µ ± ∞ ¹   +3.29% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"X|Y|Z"-4                                                          60.69µ ± ∞ ¹                             59.29µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1"-4                                                               429.6m ± ∞ ¹                             434.2m ± ∞ ¹        ~ (p=0.686 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j="foo"-4                                                       315.1m ± ∞ ¹                             308.3m ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n="1"-4                                                       305.9m ± ∞ ¹                             310.3m ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j!="foo"-4                                                      301.3m ± ∞ ¹                             301.6m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".*"-4                                                              8.570 ± ∞ ¹                              8.380 ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".+"-4                                                              17.80 ± ∞ ¹                              11.29 ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"^.+$",j=~"X.+"-4                                                  23.43m ± ∞ ¹                             23.99m ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~""-4                                                                3.058 ± ∞ ¹                              3.033 ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i!=""-4                                                                11.39 ± ∞ ¹                              11.00 ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",j="foo"-4                                               324.9m ± ∞ ¹                             322.8m ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",j="foo"-4                                             48.81µ ± ∞ ¹                             47.95µ ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",i!="2",j="foo"-4                                        336.8m ± ∞ ¹                             335.7m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!=""-4                                                          3.010 ± ∞ ¹                              2.981 ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j="foo"-4                                                  2.883 ± ∞ ¹                              2.850 ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"X.+"-4                                                23.17m ± ∞ ¹                             22.48m ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"XXX|YYY"-4                                            33.95µ ± ∞ ¹                             34.41µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                                            105.4µ ± ∞ ¹                             107.1µ ± ∞ ¹        ~ (p=0.114 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",j="foo"-4                                                2.886 ± ∞ ¹                              2.842 ± ∞ ¹   -1.53% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"1.+",j="foo"-4                                              335.9m ± ∞ ¹                             339.0m ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!="2",j="foo"-4                                         2.866 ± ∞ ¹                              2.902 ± ∞ ¹        ~ (p=0.343 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4                                       3.141 ± ∞ ¹                              3.115 ± ∞ ¹        ~ (p=0.486 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4                                47.69µ ± ∞ ¹                             47.79µ ± ∞ ¹        ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"0xxx|1xxx|2xxx"-4                                                 709.0µ ± ∞ ¹                             829.6µ ± ∞ ¹  +17.01% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"(0|1|2)xxx"-4                                                     686.2µ ± ∞ ¹                             841.7µ ± ∞ ¹  +22.67% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"[0-2]xxx"-4                                                       705.5µ ± ∞ ¹                             821.5µ ± ∞ ¹  +16.45% (p=0.029 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i!~[0-2]xxx-4                                                          8.906 ± ∞ ¹                              7.991 ± ∞ ¹        ~ (p=0.200 n=4)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".*",_i!~[0-2]xxx-4                                                 8.585 ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"<unique_prefix>.+"-4                                              267.5µ ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"<unique_prefix>.+"-4                                        3.221m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!~"<unique_prefix>.+"-4                                        421.8m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4                              32.15m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n=~".+",i=~"<unique_prefix>.+"-4                              105.3m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n=~".*",i=~"<unique_prefix>.+"-4                              27.78m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~".*",n=~".*",i=~"<unique_prefix>.+"-4                              321.6µ ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~".+",n=~".+",i=~"<unique_prefix>.+"-4                              145.8m ± ∞ ¹
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/p!=""-4                                                                1.730 ± ∞ ¹
geomean                                                                                                                     24.86m                                   23.09m         +0.13%               ²
¹ need >= 6 samples for confidence interval at level 0.95
² benchmark set differs from baseline; geomeans may not be comparable

                                                                                                  │ Clean-up-symbolization-67b7a0af7.txt │ Have-a-lower-bound-to-using-single-pass-stringify-bd654b059.txt │
                                                                                                  │                 B/op                 │                  B/op                    vs base                │
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X"-4                                                           2.166Ki ± ∞ ¹                             2.167Ki ± ∞ ¹       ~ (p=0.143 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j="foo"-4                                                   2.823Ki ± ∞ ¹                             2.823Ki ± ∞ ¹       ~ (p=1.000 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j!="foo"-4                                                  2.823Ki ± ∞ ¹                             2.823Ki ± ∞ ¹       ~ (p=0.714 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"XXX|YYY"-4                                                    1.836Ki ± ∞ ¹                             1.836Ki ± ∞ ¹       ~ (p=1.000 n=4) ²
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"X.+"-4                                                        1.781Ki ± ∞ ¹                             1.781Ki ± ∞ ¹       ~ (p=1.000 n=4) ²
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"X|Y|Z"-4                                                      2.760Ki ± ∞ ¹                             2.760Ki ± ∞ ¹       ~ (p=0.571 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1"-4                                                           145.2Mi ± ∞ ¹                             145.2Mi ± ∞ ¹       ~ (p=0.886 n=4)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j="foo"-4                                                   92.60Mi ± ∞ ¹                             92.69Mi ± ∞ ¹       ~ (p=0.686 n=4)
…